Move ShowFmt to Cardano.Wallet.Util
